Room using new grid class & cleanup

+++ LotF/src/main/java/lotf/util/math/MathH.java
@@ -6,12 +6,12 @@ public class MathH {
 	
 	public static int floor(float value) {
 		int i = (int) value;
-		return value < (float) i ? i - 1 : i;
+		return value < i ? i - 1 : i;
 	}
 	
 	public static int ceil(float value) {
 		int i = (int)value;
-		return value > (float) i ? i + 1 : i;
+		return value > i ? i + 1 : i;
 	}
 	
 	public static int fClamp(float num, float min, float max) {
